refactor: 야구, 축구는 페이지 포맷이 동일하여 추상클래스로 묶어줌
